Create a custom hook useUsers for Users functionality
